Polly v7 Upgrade
Switch to using the IAsyncPolicy and ISyncPolicy interfaces to work
with the latest polly version. This splits out the execution types
into two seeparate classes. You can no longer accidentally use an asyc
policy synchronously and vice versa.
